Refactored Existing Tests
